+++ compiler/Main.hs
@@ -72,24 +72,25 @@ cppCodegen options@Cpp {..} = do
     concurrentlyFor_ files $ codeGen options typeMapping $
         [ reflection_h
         , types_cpp
+        , types_comm_cpp
         , types_h header enum_header allocator
         , apply_h applyProto apply_attribute
         , apply_cpp applyProto
+        , comm_h
         ] <>
         [ enum_h | enum_header]
   where
     applyProto = map snd $ filter (enabled apply) protocols
     enabled a p = null a || fst p `elem` a
     protocols =
-        [ (Compact, Protocol "bond::CompactBinaryReader<bond::InputBuffer>"
-                             "bond::CompactBinaryWriter<bond::OutputBuffer>")
-        , (Fast,    Protocol "bond::FastBinaryReader<bond::InputBuffer>"
-                             "bond::FastBinaryWriter<bond::OutputBuffer>")
-        , (Simple,  Protocol "bond::SimpleBinaryReader<bond::InputBuffer>"
-                             "bond::SimpleBinaryWriter<bond::OutputBuffer>")
+        [ (Compact, Protocol " ::bond::CompactBinaryReader< ::bond::InputBuffer>"
+                             " ::bond::CompactBinaryWriter< ::bond::OutputBuffer>")
+        , (Fast,    Protocol " ::bond::FastBinaryReader< ::bond::InputBuffer>"
+                             " ::bond::FastBinaryWriter< ::bond::OutputBuffer>")
+        , (Simple,  Protocol " ::bond::SimpleBinaryReader< ::bond::InputBuffer>"
+                             " ::bond::SimpleBinaryWriter< ::bond::OutputBuffer>")
         ]
 cppCodegen _ = error "cppCodegen: impossible happened."
-
 
 csCodegen :: Options -> IO()
 csCodegen options@Cs {..} = do
@@ -99,7 +100,7 @@ csCodegen options@Cs {..} = do
                  then PublicFields
                  else Properties
     let typeMapping = if collection_interfaces then csCollectionInterfacesTypeMapping else csTypeMapping
-    let templates = [ types_cs Class fieldMapping ]
+    let templates = [ comm_interface_cs , comm_proxy_cs , comm_service_cs , types_cs Class fieldMapping ]
     concurrentlyFor_ files $ codeGen options typeMapping templates
 csCodegen _ = error "csCodegen: impossible happened."
 
@@ -117,4 +118,3 @@ codeGen options typeMapping templates file = do
         createDirectoryIfMissing True outputDir
         let content = if (no_banner options) then code else (commonHeader "//" fileName <> code)
         L.writeFile (outputDir </> fileName) content
-