+++ compiler/src/Language/Bond/Codegen/Cpp/Types_cpp.hs
@@ -14,7 +14,7 @@ import Language.Bond.Codegen.TypeMapping
 import Language.Bond.Codegen.Util
 import qualified Language.Bond.Codegen.Cpp.Util as CPP
 
--- | Codegen template for generating /base_name/_type.cpp containing
+-- | Codegen template for generating /base_name/_types.cpp containing
 -- definitions of helper functions and schema metadata static variables.
 types_cpp :: MappingContext -> String -> [Import] -> [Declaration] -> (String, Text)
 types_cpp cpp file _imports declarations = ("_types.cpp", [lt|
@@ -31,6 +31,10 @@ types_cpp cpp file _imports declarations = ("_types.cpp", [lt|
         if null declParams then CPP.schemaMetadata cpp s else mempty
 
     -- global variables for enum name/value conversions
+    --
+    -- ToString is intentionally not implemented in terms of FromEnum, as
+    -- ToString returns a reference to the name stored in the map. FromEnum
+    -- copies this name into the output paramater.
     statics Enum {..} = [lt|
     namespace _bond_enumerators
     {
@@ -43,7 +47,7 @@ types_cpp cpp file _imports declarations = ("_types.cpp", [lt|
 
         const
         std::map<enum #{declName}, std::string> _value_to_name_#{declName} =
-            bond::reverse_map(_name_to_value_#{declName});
+            ::bond::reverse_map(_name_to_value_#{declName});
 
         const std::string& ToString(enum #{declName} value)
         {
@@ -51,22 +55,16 @@ types_cpp cpp file _imports declarations = ("_types.cpp", [lt|
                 GetValueToNameMap(value).find(value);
 
             if (GetValueToNameMap(value).end() == it)
-                bond::InvalidEnumValueException(value, "#{declName}");
+                ::bond::InvalidEnumValueException(value, "#{declName}");
 
             return it->second;
         }
 
         void FromString(const std::string& name, enum #{declName}& value)
         {
-            std::map<std::string, enum #{declName}>::const_iterator it =
-                _name_to_value_#{declName}.find(name);
-
-            if (_name_to_value_#{declName}.end() == it)
-                bond::InvalidEnumValueException(name.c_str(), "#{declName}");
-
-            value = it->second;
+            if (!ToEnum(value, name))
+                ::bond::InvalidEnumValueException(name.c_str(), "#{declName}");
         }
-
     } // namespace #{declName}
     } // namespace _bond_enumerators|]
       where