+++ cpp/inc/bond/protocol/compact_binary.h
@@ -597,7 +597,7 @@ public:
     }
 
     // WriteFieldBegin
-    void WriteFieldBegin(BondDataType type, uint16_t id, const bond::Metadata& /*metadata*/)
+    void WriteFieldBegin(BondDataType type, uint16_t id, const ::bond::Metadata& /*metadata*/)
     {
         WriteFieldBegin(type, id);
     }